Request one-time position update instead of starting continuous updates.
It's better to make one call to PositionSource.update() than starting
continuous updates with PositionSource.start() and then not forgetting to
call PositionSource.stop() afterwards.

Unfortunately, there is a bug in Harmattan (QTMOBILITY-1956) and update()
method doesn't work, so Harmattan code was left to use start()/stop()
sequence. However, there is no need to both set PositionSource.active to
true and call PositionSource.start(). It is enought to do either former,
or latter.