[SYM, BB10] Delay starting search one second after user stops typing.
On Harmattan it looks like onTextChanged signal is fired only after a
whole word/space was typed in. But nn Symbian and BlackBerry it's fired
after each letter was typed. So, not to flood a server with search
requests, we're only starting a search one second after user stops
tpying/deleting the text.