HAFAS XML: Indicate that error comes form the backend
There are some error messages generated by the service that might trick
users to think that the error is reported by the client, not by the
service. For example, OBB can return "An error occured." message.

To avoid this confusion, all error messages that come from the service
are now prepended with "<servicename> replied: ".