Fix some daylight saving time issues by using UTC.
This commit switches the internal clock of the hafas provider to use
UTC. hafas already provides times that are adjusted for daylight saving
so no further adjustment needs to be done by the parser.
The patch also uses 1970-01-01 as the reference day so that a day switch
during reesult generation won't result in wrong duration displays.

This commit fixes issue #185.